dev-requirements.txt
pytest-ckan
unidecode
